pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "wrknv"
dynamic = ["version"]
description = "🧰🌍 Work Environment - Foundation for the provide.io ecosystem"
authors = [
    {name = "Provide.io", email = "engineering@provide.io"},
]
readme = "README.md"
license = "Apache-2.0"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Build Tools",
    "Topic :: System :: Installation/Setup",
]
requires-python = ">=3.11"
dependencies = [
    "click>=8.0",
    "packaging>=23.0",
    "semver>=3.0.0",
    "tomli-w>=1.0.0",
    "rich>=13.0.0",
    "jinja2>=3.0.0",
    "attrs>=23.0.0",
    "cattrs>=23.0.0",
    "provide-foundation[all]>=0.0.0",
]

[project.optional-dependencies]
package = [
    # "flavor",  # Install with: uv pip install -e ~/code/gh/provide-io/flavor
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.0",
    "black>=23.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
    "bandit>=1.7.0",
]

[project.scripts]
wrknv = "wrknv.cli:entry_point"

[project.urls]
"Homepage" = "https://github.com/livingstaccato/wrknv"
"Bug Tracker" = "https://github.com/livingstaccato/wrknv/issues"
"Documentation" = "https://github.com/livingstaccato/wrknv#readme"
"Source Code" = "https://github.com/livingstaccato/wrknv"

[tool.setuptools]
packages = [
    "wrknv",
    "wrknv.cli",
    "wrknv.wenv",
    "wrknv.wenv.managers",
    "wrknv.wenv.operations",
    "wrknv.package"
]
package-dir = {"" = "src"}

[tool.setuptools.package-data]
wrknv = ["py.typed"]

[tool.setuptools.dynamic]
version = {file = "VERSION"}

[tool.black]
line-length = 111
target-version = ['py311', 'py312', 'py313']
include = '\.pyi?$'

################################################################################
# Ruff Linter and Formatter Configuration
################################################################################

[tool.ruff]
line-length = 111
indent-width = 4
target-version = "py311"
extend-exclude = ["scraps"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "ANN", "B", "C90", "SIM", "PTH", "RUF"]
ignore = ["ANN401", "B008", "E501"]

[tool.ruff.lint.isort]
known-first-party = ["wrknv", "tests"]
force-sort-within-sections = true
combine-as-imports = true

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["ANN001", "ANN002", "ANN003", "ANN201", "ANN202", "ANN204", "E402"]

[tool.ruff.lint.mccabe]
max-complexity = 25

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

################################################################################
# MyPy Static Type Checker Configuration
################################################################################

[tool.mypy]
mypy_path = "src"
show_error_codes = true
show_column_numbers = true
pretty = true
python_version = "3.11"
warn_unused_ignores = true
warn_unused_configs = true
follow_imports = "normal"
disallow_untyped_decorators = false

[[tool.mypy.overrides]]
module = [
    "packaging.*",
    "semver.*",
    "rich.*",
    "jinja2.*",
    "cattrs.*",
    "tomli_w.*",
    "opentelemetry.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
log_cli = true
log_cli_level = "DEBUG"

asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
asyncio_default_test_loop_scope = "function"

minversion = "7.0"
testpaths = ["tests"]
pythonpath = ["src", "."]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*", "*Tests"]
python_functions = ["test_*", "*_test"]
addopts = "-m 'not integration and not slow and not network' -rFE -q --color=yes --benchmark-columns=min,max,mean,stddev,median,iqr,ops --benchmark-sort=mean --dist worksteal"
markers = [
    "unit: fast unit tests",
    "integration: integration tests requiring setup",
    "cli: CLI command tests",
    "config: configuration tests",
    "container: container management tests",
    "package: package management tests",
    "operations: file and platform operations tests",
    "managers: tool manager tests",
    "slow: tests taking >1s",
    "fast: tests taking <100ms",
    "network: tests requiring network access",
    "benchmark: performance/timing sensitive tests",
    "flaky: tests known to be intermittently failing",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore:cannot collect test class 'Test.*' because it has a __init__ constructor:pytest.PytestCollectionWarning",
    "ignore:cannot collect test class .* because it has a __init__ constructor:pytest.PytestCollectionWarning",
    "ignore:coroutine .* was never awaited:RuntimeWarning",
    "ignore:Module already imported so cannot be rewritten:pytest.PytestAssertRewriteWarning",
]
norecursedirs = [
    ".git", ".hg", ".svn", "*_build", "build", "dist", "*.egg-info",
    ".venv", "venv", "workenv",
    "htmlcov", "docs/_build",
    ".hypothesis",
]

[tool.coverage.run]
source = ["src/wrknv"]
branch = true
parallel = true
omit = ["*/tests/*", "*/__pycache__/*"]

[tool.coverage.report]
show_missing = true
skip_covered = true
fail_under = 50
precision = 2
exclude_lines = [
    "pragma: no cover",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "def __repr__",
    "def __str__",
    "@(abc\\.)?abstractmethod",
    "pass",
    "logger\\.(debug|info|warning|error|exception|critical|trace)\\(",
    "print\\(",
]

[tool.coverage.html]
directory = "htmlcov"

[tool.coverage.xml]
output = "coverage.xml"

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "pytest-benchmark>=5.1.0",
    "pytest-cov>=6.2.1",
    "pytest-xdist>=3.8.0",
]

[tool.flavor]
# Configuration for packaging wrknv with flavor
entry_point = "wrknv.cli:entry_point"
app_name = "wrknv"
description = "🧰🌍 Work Environment - Foundation for the provide.io ecosystem"
